chore: create src folder for notifications and service message
