+++ client/package.json
@@ -64,18 +64,14 @@
     "@equinor/fusion-react-tooltip": "^1.1.6",
     "@microsoft/applicationinsights-web": "^3.0.2",
     "@microsoft/signalr": "^7.0.0",
-    "adaptivecards": "^2.11.2",
     "dompurify": "^2.4.1",
     "luxon": "^3.2.1",
-    "marked": "^4.2.5",
-    "re-resizable": "^6.9.9",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-query": "^3.39.2",
     "react-router-dom": "^6.9.0",
     "rxjs": "^7.5.7",
-    "styled-components": "5.3.6",
-    "swiper": "^9.3.2"
+    "styled-components": "5.3.6"
   },
   "workspaces": [
     "packages/**"
