+++ client/packages/notification/src/components/adaptivCard/defaultHostConfig.ts
