+++ client/packages/notification/src/components/adaptivCard/styling.ts
