--- client/packages/service-message/index.ts
@@ -1,3 +1,0 @@
-export * from "./components/ServiceMessage"
-export * from "./components"
-export * from "./provider/ServiceMessageProvider"
