+++ client/packages/service-message/src/components/AppChips.tsx
