+++ client/packages/service-message/src/components/ServiceMessageCardStyles.ts
