+++ client/packages/service-message/src/components/ServiceMessageIcon.tsx
