+++ client/packages/service-message/src/index.ts
@@ -1,0 +1,3 @@
+export * from './components/ServiceMessage';
+export * from './components';
+export * from './provider/ServiceMessageProvider';
