+++ client/tsconfig.base.json
@@ -21,8 +21,8 @@
 			"@equinor/portal-pages": ["packages/portal-pages/src/index.ts"],
 			"@equinor/portal-ui": ["packages/portal-ui/src/index.ts"],
 			"@equinor/portal-utils": ["packages/portal-utils/src/index.ts"],
-			"@equinor/service-message": ["packages/service-message/index.ts"],
-			"@equinor/notification": ["packages/notification/index.ts"]
+			"@equinor/service-message": ["packages/service-message/src/index.ts"],
+			"@equinor/notification": ["packages/notification/src/index.ts"]
 		}
 	},
 	"include": ["dompurify.d.ts"],
